Renamed the admin API file
