+++ src/app/admin/components/adminBlocks/adminBlocks.component.ts
@@ -40,7 +40,7 @@ import { AdminService } from "@app/services/admin.service";
 import { AlertsService } from "@app/services/alerts.service";
 import { ApiClientService } from "@app/services/apiClient.service";
 import { type BlockedUser } from "@app/interfaces/user.interface";
-import { type BlockUserResponse } from "@app/interfaces/responses";
+import { type BlockUserResponse } from "@app/interfaces/api/admin";
 
 @Component({
   selector: "app-admin-blocks",
