+++ src/app/components/forms/displayNameEditForm/displayNameEditForm.component.ts
@@ -42,7 +42,7 @@ import { ValidationService } from "@app/services/validation.service";
 import { AlertsService } from "@app/services/alerts.service";
 import { type ReportData } from "@app/interfaces/report.interface";
 import { type PartialUser } from "@app/interfaces/user.interface";
-import { UpdatedUserReportResponse } from "@app/interfaces/responses";
+import { UpdatedUserReportResponse } from "@app/interfaces/api/admin";
 import { PopUp } from "@common/popUp/popUp.component";
 import { TeleportDirective } from "@app/directives/teleport.directive";
